vehicle/requirements.txt
Adafruit_ADS1x15
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-mcp230xx
click
coloredlogs
csaps
evdev
geopy==1.22.0
netifaces
odrive==0.4.12
psutil
python-socketio[asyncio_client]
pyyaml
scipy
svgpathtools
websocket-client
zmq
